+++ apps/web/src/apis/postCreateTournament.ts
@@ -1,11 +1,10 @@
 import { clientApi } from '@/apis/client';
 import { ENDPOINTS } from '@/consts/api';
 import type { ApiResponseT } from '@/types/api';
-
 import type {
   PostCreateTournamentRequestT,
   PostCreateTournamentResponseT,
-} from '../_types/tournament';
+} from '@/types/tournament';
 
 export const postCreateTournament = async (body: PostCreateTournamentRequestT) => {
   const { data } = await clientApi.post<ApiResponseT<PostCreateTournamentResponseT>>(
