+++ apps/web/src/app/archive/tournament/_components/TournamentFab.tsx
@@ -1,17 +1,24 @@
 import { AddIconFill } from '@/assets/icons';
 import Button from '@/components/button';
+import CreateTournamentDialogContent from '@/components/common/create-tournament-dialog';
+import { Dialog, DialogTrigger } from '@/components/dialog';
 
 function TournamentFab() {
   return (
-    <Button
-      variant="primary"
-      size="xl"
-      icon="only"
-      aria-label="토너먼트 만들기"
-      className="fixed right-[max(20px,calc(50%-480px/2+20px))] bottom-[103px] z-30"
-    >
-      <AddIconFill width={30} height={30} />
-    </Button>
+    <Dialog>
+      <DialogTrigger asChild>
+        <Button
+          variant="primary"
+          size="xl"
+          icon="only"
+          aria-label="토너먼트 만들기"
+          className="fixed right-[max(20px,calc(50%-480px/2+20px))] bottom-[103px] z-30"
+        >
+          <AddIconFill width={30} height={30} />
+        </Button>
+      </DialogTrigger>
+      <CreateTournamentDialogContent />
+    </Dialog>
   );
 }
